Fixed due checking bug
If the item is constructed, then immediately tested to see if it's due,
it will report as due (even if it hasn't been checked out yet) because
the default value for a DateTime (and, by extension, DueDate) is over
2,000 years in the past.

Also, at a rate of 10 cents per day, that would be about $73,000 in late
fees. I'd hate to be paying *those* fines.